% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as.FLStock,FLBiol-method}
\alias{as.FLStock,FLBiol-method}
\title{Create an \code{FLStock} object from \code{FLBiol} and \code{FLFishery} or \code{FLFisheries}}
\usage{
\S4method{as.FLStock}{FLBiol}(
  object,
  fisheries,
  full = TRUE,
  catch = rep(1, length(fisheries)),
  ...
)
}
\arguments{
\item{object}{An object of class \code{FLBiol}.}

\item{fisheries}{An object of classes \code{FLFishery} or \code{FLFisheries}.}

\item{full}{Should fishing mortality be computed and added? Logical.}

\item{catch}{Vector of the same length ad \code{fisheries}, indicating the
position of the \code{FLCatch} objects that refer to the \code{FLBiol}. Defaults to
the first one along all elements.}
}
\value{
An object of class \code{FLStock}.
}
\description{
A coertion method that returns a \code{FLStock} object by combining biological
information contained in \code{FLBiol} and catch data from one of more
\code{FLFishery} objects.
}
\details{
Details: Aliquam sagittis feugiat felis eget consequat.
}
\examples{
data(nsfishery)
as.FLStock(sol, nsfleet)
}
\seealso{
FLCore::FLStock
}
\author{
The FLR Team
}
\keyword{methods}
