% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcb.R
\name{FCB,ANY-method}
\alias{FCB,ANY-method}
\alias{FCB,missing-method}
\alias{FCB,FLBiols-method}
\alias{FCB,FLFisheries-method}
\title{Matrix of Fishery - Catch - Biol relationships}
\usage{
\S4method{FCB}{ANY}(object, ...)

\S4method{FCB}{missing}(object, ...)

\S4method{FCB}{FLBiols}(object, fisheries)

\S4method{FCB}{FLFisheries}(object, biols)
}
\arguments{
\item{object}{A list or vector containing a row for the matrix.}

\item{...}{Further vectors to be merged into the matrix.}

\item{fisheries}{An `FLFisheries` object to extract fishery and catch names from.}

\item{biols}{An `FLBiols` object to match with `FLCatch` elements in `object`.}
}
\description{
The relationships between a fishery, its catch elements and the biological
populations that catch is taken from, is part of the `fwdControl` class. When
a single `FLBiol` and `FLFishery`, or just an `FLStock`, are projected
forward, this structure is constructed on the fly. Even when multiple biols
and fisheries are used, a guess is made based on name matching.
}
\details{
But when a more complex structure is employed, the `FCB` matrix can be given
to the `fwdControl()` constructor method. Elements in this matrix can be
names or numbers.

If `FLBiols` and `FLFisheries` objects are passed to `FCB`, a guess is made
at constructing the matrix based on the names of the various list elements.
}
\examples{
# 1 fishery with catches from 2 biols
FCB(c(f=1, c=1, b=2), c(f=1, c=2, b=2))
# 2 fisheries with caches from 3 biols
FCB(c(f=1, c=1, b=1), c(f=1, c=2, b=2),
  c(f=2, c=1, b=2), c(f=2, c=2, b=2),
  c(f=2, c=3, b=3))
}
