% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/classes.R, R/constructors.R
\docType{class}
\name{FLCatches}
\alias{FLCatches}
\alias{FLCatches-methods}
\alias{FLCatches}
\alias{FLCatches-class}
\alias{FLCatches-methods}
\alias{FLCatches,list-method}
\alias{FLCatches,missing-method}
\title{FLCatches constructor}
\usage{
FLCatches(object, ...)

\S4method{FLCatches}{list}(object)

\S4method{FLCatches}{missing}(object, ...)
}
\arguments{
\item{object}{Either a list of FLCatch objects or missing}

\item{...}{Other things}
}
\value{
An FLCatches object
}
\description{
Make an FLCatches object.

This is class is used inside \code{FLFishery} to store the catches of
all species caught by that fleet. It is not meant to be used directly.
}
\details{
Make an FLCatches object.
}
\section{Validity}{


  \describe{
    \item{Length of dimensions 2:5}{All elements must be of class \code{FLCatch}}
    \item{Length of dimensions 2, 4 and 5}{All \code{FLQuant} slots must share
      dimensions 2, 4 and 5 (year, season and area).}
    \item{quant must 'age'}{The 1st dimension in elements must be 'age'.}
  }

You can inspect the class validity function by using
   \code{getValidity(getClassDef('FLCatches'))}
}

\section{Accessors}{

Elements in the classes can be extracted and replaced using the list subset
operators,'[', '[<-', '[[' and '[[<-'.

The values passed for replacement need to be of the class FLCatch.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its elements.
}

\section{Methods}{

Methods exist for various operations with elements stored in the class:

\describe{
 \item{plot}{Standard plot for the FLCatches class.}
}
}

\seealso{
\link{FLComp}

\link{FLCatch}, \link{FLFishery}
}
\author{
The FLR Team

Iago Mosqueira, EC JRC.
}
\keyword{classes}
