% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLFishery.R
\docType{package}
\name{FLFishery}
\alias{FLFishery}
\alias{FLFishery-package}
\title{Classes and Methods for Modeling of Fishing Fleets}
\description{
FLFishery defines a set of classes to load, manipulate and combine landings
discards data from fishing fleets catching individuals of one or more stocks.
}
\examples{
\dontrun{FLFishery()}
}
\author{
Iago Mosqueira (EC JRC) \email{iago.mosqueira@jrc.ec.europa.eu}

Maintainer: Iago Mosqueira (EC JRC) \email{iago.mosqueira@jrc.ec.europa.eu}
}
\seealso{
See \code{vignette("FLFishery", package = "FLFishery")} for an
overview of the package.
}

