% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/accessors.R
\docType{class}
\name{FLFisherycpp}
\alias{FLFisherycpp}
\alias{FLFisherycpp-methods}
\alias{FLFisherycpp-class}
\alias{vcost,FLFisherycpp-method}
\alias{vcost<-,FLFisherycpp,FLQuant-method}
\alias{vcost<-,FLFisherycpp,numeric-method}
\alias{crewshare,FLFisherycpp-method}
\alias{crewshare<-,FLFisherycpp,FLQuant-method}
\alias{crewshare<-,FLFisherycpp,numeric-method}
\alias{hperiod,FLFisherycpp-method}
\alias{hperiod<-,FLFisherycpp,FLQuant-method}
\alias{hperiod<-,FLFisherycpp,numeric-method}
\alias{fcost<-,FLFisherycpp,FLQuant-method}
\alias{fcost<-,FLFisherycpp,numeric-method}
\alias{fcost,FLFisherycpp-method}
\alias{orevenue<-,FLFisherycpp,FLQuant-method}
\alias{orevenue<-,FLFisherycpp,numeric-method}
\alias{orevenue,FLFisherycpp-method}
\alias{effort,FLFisherycpp,ANY-method}
\alias{effort<-,FLFisherycpp,FLQuant-method}
\alias{effort<-,FLFisherycpp,numeric-method}
\alias{capacity,FLFisherycpp-method}
\alias{capacity<-,FLFisherycpp,FLQuant-method}
\alias{capacity<-,FLFisherycpp,numeric-method}
\alias{hperiod,FLFisherycpp-method}
\alias{hperiod<-,FLFisherycpp,FLQuant-method}
\alias{hperiod<-,FLFisherycpp,numeric-method}
\alias{orevenue,FLFisherycpp-method}
\alias{orevenue<-,FLFisherycpp,FLQuant-method}
\alias{orevenue<-,FLFisherycpp,numeric-method}
\alias{effort,FLFisherycpp,ANY-method}
\alias{effort<-,FLFisherycpp,FLQuant-method}
\alias{effort<-,FLFisherycpp,numeric-method}
\alias{vcost,FLFisherycpp-method}
\alias{vcost<-,FLFisherycpp,FLQuant-method}
\alias{vcost<-,FLFisherycpp,numeric-method}
\alias{fcost,FLFisherycpp-method}
\alias{fcost<-,FLFisherycpp,FLQuant-method}
\alias{fcost<-,FLFisherycpp,numeric-method}
\alias{crewshare,FLFisherycpp-method}
\alias{crewshare<-,FLFisherycpp,FLQuant-method}
\alias{crewshare<-,FLFisherycpp,numeric-method}
\title{An internal class for homogeneous fishing fleets}
\usage{
\S4method{capacity}{FLFisherycpp}(object)

\S4method{capacity}{FLFisherycpp,FLQuant}(object) <- value

\S4method{capacity}{FLFisherycpp,numeric}(object) <- value

\S4method{hperiod}{FLFisherycpp}(object)

\S4method{hperiod}{FLFisherycpp,FLQuant}(object) <- value

\S4method{hperiod}{FLFisherycpp,numeric}(object) <- value

\S4method{orevenue}{FLFisherycpp}(object)

\S4method{orevenue}{FLFisherycpp,FLQuant}(object) <- value

\S4method{orevenue}{FLFisherycpp,numeric}(object) <- value

\S4method{effort}{FLFisherycpp,ANY}(object)

\S4method{effort}{FLFisherycpp,FLQuant}(object) <- value

\S4method{effort}{FLFisherycpp,numeric}(object) <- value

\S4method{vcost}{FLFisherycpp}(object)

\S4method{vcost}{FLFisherycpp,FLQuant}(object) <- value

\S4method{vcost}{FLFisherycpp,numeric}(object) <- value

\S4method{fcost}{FLFisherycpp}(object)

\S4method{fcost}{FLFisherycpp,FLQuant}(object) <- value

\S4method{fcost}{FLFisherycpp,numeric}(object) <- value

\S4method{crewshare}{FLFisherycpp}(object)

\S4method{crewshare}{FLFisherycpp,FLQuant}(object) <- value

\S4method{crewshare}{FLFisherycpp,numeric}(object) <- value
}
\arguments{
\item{object}{Object to extract or operate on}

\item{value}{Replacement value}
}
\description{
The same as the FLFishery class but all predictModel slots have been turned into FLQuant objects.
The class is used for passing to C++ FLR objects.
}
\details{
What have you done what's in your mind what do you need
Where shall we go to let it out
What have you seen, we don't know where you've been
Life so often blows your candle out
Believe in what is right, what's right for you tonight
Who knows what the fuck it's all about
}
\section{Slots}{


\code{FLFisherycpp} objects inherit from \code{FLCatches} woth a number of slots
added.
    \describe{
    \item{.Data}{The list of \code{FLCatch} object with catch data per stock,
      (\code{FLCatches}).}
    \item{name}{Name of the object, e.g. species or stock code, (\code{character}).}
    \item{desc}{Description of the data contents and origin, (\code{character}).}
    \item{range}{Ranges of age and years, plusgroup, (\code{numeric}).}
    \item{capacity}{Number of boats in the fleet, (\code{FLQuant}).}
    \item{effort}{Mean effort per boat applied by the fleet, (\code{FLQuant}).}
    \item{hperiod}{Start and end of fishing within each time step, as
      proportions. An \code{FLQuant} object with dimnames
      `quant=c('start', 'end')` in the first dimension.}
    \item{vcost}{Variable costs per unit of effort, (\code{FLQuant}).}
    \item{fcost}{Variable costs per unit of effort, (\code{FLQuant}).}
    \item{orevenue}{Revenues obtained from sources other than landings, (\code{FLQuant}).}
    \item{crewshare}{Formula, parameter values and inputs to calculate the
      crew costs, (\code{FLQuant}).}
}
}

\section{Validity}{


You can inspect the class validity function by using
   \code{getValidity(getClassDef('FLFishery'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:

\describe{
  \item{ccost}{Calculate the total crew costs by evaluating the formula in
    \code{crewshare}.}
  \item{cost}{Total costs, calculated.}
  \item{lrevenue}{.}
  \item{revenue}{.}
  \item{profit}{.}
  \item{landings}{.}
  \item{discards}{.}
  \item{catch}{.}
  \item{catch.n}{.}
  \item{catch.wt}{.}
  \item{harvest}{.}
}
}

\examples{

data(ple4)
FLFishery(PLE=as(ple4, 'FLCatch'))
}
\seealso{
\link{FLCatches}
}
\author{
Iago Mosqueira, EC JRC.
}
\keyword{classes}
