% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsfishery}
\alias{nsfishery}
\alias{nsfleet}
\alias{ple}
\alias{sol}
\title{Simulated fisheries data for fisheries exploiting North sea plaice and sole.}
\format{
The objects of classes FLFisheries (nsfleet) and FLBiol (sol, ple).

An object of class \code{FLBiol} of length 1.

An object of class \code{FLBiol} of length 1.
}
\usage{
nsfleet

ple

sol
}
\description{
An example data for two fisheries (*bt* and *gn*) exploiting two stocks,
*ple* and *sol*, represented by two objects of class *FLBiol*.
}
\keyword{datasets}
