% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/classes.R, R/accessors.R,
%   R/computation.R, R/constructors.R, R/methods.R, R/plot.R
\docType{class}
\name{FLCatch}
\alias{FLCatch}
\alias{FLCatch-method}
\alias{catch.sel<-}
\alias{catch.sel<--method}
\alias{discards.ratio}
\alias{discards.ratio-method}
\alias{lrevenue}
\alias{lrevenue-method}
\alias{FLCatch-class}
\alias{FLCatch-methods}
\alias{landings.n,FLCatch-method}
\alias{name,FLCatch-method}
\alias{desc,FLCatch-method}
\alias{range,FLCatch-method}
\alias{landings.wt,FLCatch-method}
\alias{discards.n,FLCatch-method}
\alias{discards.wt,FLCatch-method}
\alias{catch.sel,FLCatch-method}
\alias{price,FLCatch-method}
\alias{catch.q,FLCatch-method}
\alias{name<-,FLCatch,character-method}
\alias{desc<-,FLCatch,character-method}
\alias{range<-,FLCatch,numeric-method}
\alias{landings.n<-,FLCatch,FLQuant-method}
\alias{landings.wt<-,FLCatch,FLQuant-method}
\alias{discards.n<-,FLCatch,FLQuant-method}
\alias{discards.wt<-,FLCatch,FLQuant-method}
\alias{catch.sel<-,FLCatch,FLQuant-method}
\alias{price<-,FLCatch,FLQuant-method}
\alias{catch.q<-,FLCatch,FLPar-method}
\alias{discards.sel,FLCatch-method}
\alias{landings.sel,FLCatch-method}
\alias{discards,FLCatch-method}
\alias{catch,FLCatch-method}
\alias{catch.n,FLCatch-method}
\alias{catch.wt,FLCatch-method}
\alias{landings,FLCatch-method}
\alias{discards.ratio,FLCatch-method}
\alias{landings.n<-,FLCatch,numeric-method}
\alias{landings.wt<-,FLCatch,numeric-method}
\alias{discards.n<-,FLCatch,numeric-method}
\alias{discards.wt<-,FLCatch,numeric-method}
\alias{catch.sel<-,FLCatch,numeric-method}
\alias{catch.q<-,FLCatch,numeric-method}
\alias{price<-,FLCatch,numeric-method}
\alias{FLCatch,FLQuant-method}
\alias{FLCatch,missing-method}
\alias{lrevenue,FLCatch-method}
\alias{plot,FLCatch,missing-method}
\title{FLCatch constructor}
\usage{
FLCatch(object, ...)

catch.sel(object, ...) <- value

discards.ratio(object, ...)

lrevenue(object, ...)

\S4method{landings.n}{FLCatch}(object)

\S4method{landings.n}{FLCatch,numeric}(object) <- value

\S4method{landings.wt}{FLCatch,numeric}(object) <- value

\S4method{discards.n}{FLCatch,numeric}(object) <- value

\S4method{discards.wt}{FLCatch,numeric}(object) <- value

\S4method{catch.sel}{FLCatch,numeric}(object) <- value

\S4method{price}{FLCatch,numeric}(object) <- value

\S4method{catch.q}{FLCatch,numeric}(object) <- value

\S4method{landings}{FLCatch}(object)

\S4method{discards}{FLCatch}(object)

\S4method{catch}{FLCatch}(object)

\S4method{catch.n}{FLCatch}(object)

\S4method{catch.wt}{FLCatch}(object)

\S4method{catch.wt}{FLCatch}(object)

\S4method{FLCatch}{FLQuant}(object, ...)

\S4method{FLCatch}{missing}(object, ...)

\S4method{lrevenue}{FLCatch}(object)

\S4method{landings.sel}{FLCatch}(object)

\S4method{discards.sel}{FLCatch}(object)

\S4method{discards.ratio}{FLCatch}(object)

\S4method{plot}{FLCatch,missing}(x, y, ...)
}
\arguments{
\item{object}{Either an FLQuant (to determine the size of the FLQuant slots) or missing}

\item{...}{Other things}

\item{value}{Replacement value}

\item{x}{FLCatch}

\item{y}{missing}
}
\value{
An FLCatch object
}
\description{
Make an FLCatch object.

Catch data for a single species or stock unit is handled by the
\code{FLCatch} class. Data is separated as landings and discards by age, in
numbers, with the corresponding mean weights at age.
}
\details{
Make an FLCatch object.

This is class is used inside \code{FLFishery} to store the catches of
a single stock or species caught by that fleet.
}
\section{Slots}{

    \describe{
    \item{catch.q}{Parameters of the catchability function, (\code{FLPar}).}
    \item{catch.sel}{Selectivity at age as proportions over fully-selected
      ages, (\code{FLQuant}).}
    \item{desc}{Description of the data contents and origin, (\code{character}).}
    \item{discards.n}{Discards at age in numbers, (\code{FLQuant}).}
    \item{discards.wt}{Mean weight-at-age in the discards, (\code{FLQuant}).}
    \item{name}{Name of the object, e.g. species or stock code, (\code{character}).}
    \item{landings.n}{Landings at age in numbers, (\code{FLQuant}).}
    \item{landings.wt}{Mean weight-at-age in the landings, (\code{FLQuant}).}
    \item{price}{Mean price by age per unit of weight, (\code{FLQuant}).}
    \item{range}{Ranges of age and years, plusgroup, (\code{numeric}).}
}
}

\section{Validity}{


  \describe{
    \item{Length of dimensions 2:5}{All \code{FLQuant} slots must share
      dimensions 2 to 5.}
    \item{\code{iter} dim of length 1 or N}{The 6th dimension in all
      \code{FLQuant} and \code{FLPar} slots must be 1 or N, where N is the
      same value for the whole object.}
    \item{Length of dimensions 2:5}{All \code{FLQuant} slots must share
      dimensions 2 to 5.}
  }

You can inspect the class validity function by using
   \code{getValidity(getClassDef('FLCatch'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:

\describe{
 \item{landings}{Total landings as sum on 'age' of \code{landings.n}
   times \code{landings.wt}.}
 \item{discards}{Total discards as sum on 'age' of \code{discards.n}
   times \code{discards.wt}.}
 \item{landings.sel}{Selectivity at age in the landings as proportions
     over fully-selected ages, (\code{FLQuant}).}
 \item{discards.sel}{Selectivity at age in the discards as proportions
     over fully-selected ages, (\code{FLQuant}).}
 \item{catch.n}{Catch at age in numbers as \code{landings.n} plus
   \code{discards.n}.}
 \item{catch.wt}{Weighted average of \code{landings.wt} and
   \code{discards.wt}.}
 \item{catch}{Total catch as sum of \code{landings} and
   \code{discards}.}
 \item{discards.ratio}{Proportion at age of discards in catch.}
 \item{plot}{Standard plot for the FLCatch class.}
}
}

\examples{

data(ple4)

# EXTRACT data from FLCore ple4, fake prices
fca <- FLCatch(name='PLE', desc='All NS PLE catches',
  landings.n=landings.n(ple4), landings.wt=landings.wt(ple4),
  discards.n=discards.n(ple4), discards.wt=discards.wt(ple4),
  price=landings.wt(ple4) * 23, catch.q=FLPar(q=1),
  catch.sel=catch.sel(ple4))

# Calculations
landings(fca)

catch.n(fca)
catch.wt(fca)

}
\seealso{
\link{FLComp}

\link{FLCatches}, \link{FLFishery}
}
\author{
The FLR Team

Iago Mosqueira, EC JRC.
}
\keyword{classes}
