% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/classes.R, R/accessors.R,
%   R/constructors.R
\docType{class}
\name{FLFisheries}
\alias{FLFisheries}
\alias{FLFisheries-methods}
\alias{FLFisheries}
\alias{FLFisheries-class}
\alias{FLFisheries-methods}
\alias{landings.n,FLFisheries-method}
\alias{FLFisheries,list-method}
\alias{FLFisheries,missing-method}
\title{FLFisheries constructor}
\usage{
FLFisheries(object, ...)

\S4method{landings.n}{FLFisheries}(object)

\S4method{FLFisheries}{list}(object, desc = character(1))

\S4method{FLFisheries}{missing}(object, ...)
}
\arguments{
\item{object}{Either a list of FLFishery objects or missing}

\item{...}{Other things}

\item{desc}{Description}
}
\value{
An FLFisheries object
}
\description{
Make an FLFisheries object.

This is a containuer class for \code{FLFishery} objects.
}
\details{
Make an FLFisheries object.
}
\section{Accessors}{

Elements in the classes can be extracted and replaced using the list subset
operators,'[', '[<-', '[[' and '[[<-'.

The values passed for replacement need to be of the class FLFishery.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its elements.
}

\seealso{
\link{FLComp}

\link{FLCatch}, \link{FLFishery}
}
\author{
The FLR Team

Finlay Scott, EC JRC.
}
\keyword{classes}
