% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLFishery.R, R/classes.R
\docType{package}
\name{FLFishery}
\alias{FLFishery}
\alias{FLFishery-package}
\alias{FLFishery}
\alias{FLFishery-methods}
\alias{FLFishery-class}
\title{Classes and Methods for Modeling of Fishing Fleets}
\description{
FLFishery defines a set of classes to load, manipulate and combine landings
discards data from fishing fleets catching individuals of one or more stocks.

Fishing fleets consisting of a number of boats operating homogeneously can be
modelled using the \code{FLFishery} class. All boats in the fleet must have a
common gear configuration during each time step and area (no \emph{metiers}).
}
\details{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique. 
Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id
}
\section{Slots}{


\code{FLFishery} objects inherit from \code{FLCatches} woth a number of slots
added.
    \describe{
    \item{.Data}{The list of \code{FLCatch} object with catch data per stock,
      (\code{FLCatches}).}
    \item{name}{Name of the object, e.g. species or stock code, (\code{character}).}
    \item{desc}{Description of the data contents and origin, (\code{character}).}
    \item{range}{Ranges of age and years, plusgroup, (\code{numeric}).}
    \item{capacity}{Number of boats in the fleet, (\code{FLQuant}).}
    \item{effort}{Mean effort per boat applied by the fleet, (\code{FLQuant}).}
    \item{hperiod}{Start and end of fishing within each time step, as
      proportions. An \code{FLQuant} object with dimnames
      `quant=c('start', 'end')` in the first dimension.}
    \item{vcost}{Variable costs per unit of effort, (\code{FLQuant}).}
    \item{orevenue}{Revenues obtained from sources other than landings, (\code{FLQuant}).}
    \item{crewshare}{Formula, parameter values and inputs to calculate the
      crew costs, (\code{predictModel}).}
}
}

\section{Validity}{


  \describe{
    \item{VALIDITY}{Neque porro quisquam est qui dolorem ipsum.}
}

You can inspect the class validity function by using
   \code{getValidity(getClassDef('FLFishery'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:

\describe{
  \item{ccost}{Calculate the total crew costs by evaluating the formula in
    \code{crewshare}.}
  \item{cost}{Total costs, calculated.}
  \item{lrevenue}{.}
  \item{revenue}{.}
  \item{profit}{.}
  \item{landings}{.}
  \item{discards}{.}
  \item{catch}{.}
  \item{catch.n}{.}
  \item{catch.wt}{.}
  \item{harvest}{.}
}
}

\examples{
\dontrun{FLFishery()}

data(ple4)
FLFishery(PLE=as(ple4, 'FLCatch'))
}
\seealso{
See \code{vignette("FLFishery", package = "FLFishery")} for an
overview of the package.

\link{FLCatches}
}
\author{
Iago Mosqueira (EC JRC) \email{iago.mosqueira@ec.europa.eu}

Maintainer: Iago Mosqueira (EC JRC) \email{iago.mosqueira@ec.europa.eu}

Iago Mosqueira, EC JRC.
}
\keyword{classes}
