% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLFishery.R, R/generics.R, R/classes.R,
%   R/constructors.R
\docType{package}
\name{FLFishery}
\alias{FLFishery}
\alias{FLFishery-package}
\alias{crewshare}
\alias{crewshare-method}
\alias{crewshare<-}
\alias{crewshare<--method}
\alias{hperiod}
\alias{hperiod-method}
\alias{hperiod<-}
\alias{hperiod<--method}
\alias{orevenue}
\alias{orevenue-method}
\alias{orevenue<-}
\alias{orevenue<--method}
\alias{FLFishery}
\alias{FLFishery-methods}
\alias{FLFishery-class}
\alias{vcost,FLFishery-method}
\alias{vcost<-,FLFishery,FLQuant-method}
\alias{catch.n,FLFishery-method}
\alias{catch,FLFishery-method}
\alias{ccost,FLFishery-method}
\alias{cost,FLFishery-method}
\alias{crewshare,FLFishery-method}
\alias{crewshare<-,FLFishery,predictModel-method}
\alias{profit,FLFishery-method}
\alias{orevenue,FLFishery-method}
\alias{orevenue<-,FLFishery,FLQuant-method}
\alias{orevenue<-,FLFishery,numeric-method}
\alias{lrevenue,FLFishery-method}
\alias{landings,FLFishery-method}
\alias{hperiod,FLFishery-method}
\alias{hperiod<-,FLFishery,FLQuant-method}
\alias{hperiod<-,FLFishery,numeric-method}
\alias{fcost<-,FLFishery,FLQuant-method}
\alias{fcost,FLFishery-method}
\alias{effort<-,FLFishery,FLQuant-method}
\alias{effort,FLFishery,ANY-method}
\alias{discards,FLFishery-method}
\alias{[[<-,FLFishery,character,missing,FLCatch-method}
\alias{[[<-,FLFishery,numeric,missing,FLCatch-method}
\alias{[,FLFishery,ANY,missing,ANY-method}
\alias{FLFishery,list-method}
\alias{FLFishery,FLCatch-method}
\alias{FLFishery,missing-method}
\title{Classes and Methods for Modeling of Fishing Fleets}
\usage{
crewshare(object, ...)

crewshare(object, ...) <- value

hperiod(object, ...)

hperiod(object, ...) <- value

orevenue(object, ...)

orevenue(object, ...) <- value

\S4method{FLFishery}{list}(object, ...)

\S4method{FLFishery}{FLCatch}(object, ...)

\S4method{FLFishery}{missing}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other things}

\item{value}{Replacement value}
}
\description{
FLFishery defines a set of classes to load, manipulate and combine landings
discards data from fishing fleets catching individuals of one or more stocks.

Fishing fleets consisting of a number of boats operating homogeneously can be
modelled using the \code{FLFishery} class. All boats in the fleet must have a
common gear configuration during each time step and area (no \emph{metiers}).
}
\details{
What do you say what I can ever do for you
What are we gonna do to pass the time
What do you care when you find that life's unfair
Equality is just a state of mind
Believe whatever is right what's right for you tonight
You know where to draw the line
}
\section{Slots}{


\code{FLFishery} objects inherit from \code{FLCatches} woth a number of slots
added.
    \describe{
    \item{.Data}{The list of \code{FLCatch} object with catch data per stock,
      (\code{FLCatches}).}
    \item{name}{Name of the object, e.g. species or stock code, (\code{character}).}
    \item{desc}{Description of the data contents and origin, (\code{character}).}
    \item{range}{Ranges of age and years, plusgroup, (\code{numeric}).}
    \item{capacity}{Number of boats in the fleet, (\code{FLQuant}).}
    \item{effort}{Mean effort per boat applied by the fleet, (\code{FLQuant}).}
    \item{hperiod}{Start and end of fishing within each time step, as
      proportions. An \code{FLQuant} object with dimnames
      `quant=c('start', 'end')` in the first dimension.}
    \item{vcost}{Variable costs per unit of effort, (\code{FLQuant}).}
    \item{fcost}{Variable costs per unit of effort, (\code{FLQuant}).}
    \item{orevenue}{Revenues obtained from sources other than landings, (\code{FLQuant}).}
    \item{crewshare}{Formula, parameter values and inputs to calculate the
      crew costs, (\code{predictModel}).}
}
}

\section{Validity}{


You can inspect the class validity function by using
   \code{getValidity(getClassDef('FLFishery'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:

\describe{
  \item{ccost}{Calculate the total crew costs by evaluating the formula in
    \code{crewshare}.}
  \item{cost}{Total costs, calculated.}
  \item{lrevenue}{.}
  \item{revenue}{.}
  \item{profit}{.}
  \item{landings}{.}
  \item{discards}{.}
  \item{catch}{.}
  \item{catch.n}{.}
  \item{catch.wt}{.}
  \item{harvest}{.}
}
}

\examples{
\dontrun{FLFishery()}

data(ple4)
FLFishery(PLE=as(ple4, 'FLCatch'))
}
\seealso{
See \code{vignette("FLFishery", package = "FLFishery")} for an
overview of the package.

\link{FLCatches}
}
\author{
Iago Mosqueira (EC JRC) \email{iago.mosqueira@ec.europa.eu}

Maintainer: Iago Mosqueira (EC JRC) \email{iago.mosqueira@ec.europa.eu}

Iago Mosqueira, EC JRC.
}
\keyword{classes}
