% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/harvest.R
\name{harvests}
\alias{harvests}
\alias{harvests-method}
\alias{harvest}
\alias{harvests,FLBiol,FLFisheries-method}
\alias{harvest,FLBiol,FLFishery-method}
\alias{harvest,FLBiol,FLFisheries-method}
\alias{harvest,FLBiol,FLCatch-method}
\alias{fbar,FLBiol-method}
\title{Methods to calculate fishing mortalities}
\usage{
harvests(object, catches, ...)

\S4method{harvests}{FLBiol,FLFisheries}(object, catches, fcb = rep(1, length(catches)), units = c("f", "hr"))

\S4method{harvest}{FLBiol,FLFishery}(object, catch, fcb = 1)

\S4method{harvest}{FLBiol,FLFisheries}(object, catch, fcb = 1)

\S4method{harvest}{FLBiol,FLCatch}(object, catch)

\S4method{fbar}{FLBiol}(
  object,
  fisheries,
  range = unlist(dims(object)[c("min", "max")]),
  minfbar = range[1],
  maxfbar = range[2]
)
}
\arguments{
\item{object}{Object containing population abundances, of class \emph{FLBiol}.}

\item{catches}{Object containing catches in number of the population represented by object, class \emph{FLFisheries}.}

\item{...}{Other things}

\item{fcb}{A vector indicating the correspondance, by position of name, between \emph{object} and the \emph{FLCatch} elements inside \emph{catches}, and of the same length.}

\item{units}{Should output be in terms of fishing mortaloty ('f') or harvest rate ('hr').}

\item{catch}{Object containing catches in number of the population represented by object, class \emph{FLFishery}.}
}
\value{
An object of class \emph{FLQuant} or \emph{FLQuants}.
}
\description{
Fishing mortalities and harvest rates can be calculated for a combination of
FLBiol and FLFishery/FLFisheries using the harvest() method.
}
\details{
The calculated fishing mortalities, or havest rates, are returned, in the
case of \emph{harvest}, disaggregated by fishery, as an \emph{FLQuants} list. For a
single \emph{FLFishery} object, a single \emph{FLQuant} is obtained.
}
\examples{
data(nsfishery)
harvests(ple, nsfleet)
harvest(ple, nsfleet[["bt"]], fcb="ple")
harvest(ple, nsfleet)
harvest(ple, nsfleet[["bt"]][["ple"]])
fbar(ple, fisheries=nsfleet, minfbar=3, maxfbar=6)
fbar(ple, fisheries=nsfleet[["bt"]], minfbar=3, maxfbar=6)
}
\seealso{
\code{\link[FLCore:accessors]{FLCore::harvest()}}
}
\author{
The FLR Team
}
\keyword{methods}
